Added: - Changed gameobject management to guid based. - Added a helper function to get an array of components from a node tree - Camera position is now saved in the level data - Updated the editor settings window - Editor settings are now saved with the new YAML api - Added mouse panning and scrolling
Fixed:
- Fixed components not being loaded in the correct position
- Fixed components not being loaded as a root component
- Fixed components that are handled by an actor being saved
- Fixed components spawned from the spawn window not being registered

+++ Flow/Source/Flow/Maths/Maths.cpp
@@ -92,5 +92,5 @@ FLOW_API Vector3::Vector3(IntVector2 v)
 {}
 
 FLOW_API Vector4::Vector4(IntVector4 v)
-	: x(static_cast<int>(v.x)), y(static_cast<int>(v.z)), z(static_cast<int>(v.y)), w(static_cast<int>(v.w))
+	: x(static_cast<float>(v.x)), y(static_cast<float>(v.z)), z(static_cast<float>(v.y)), w(static_cast<float>(v.w))
 {}
